Refactor: LinkController를 LinkController와 LinkRestController로 분리
